streamlit/requirements.txt
bcrypt==4.0.1
numpy==1.23.5
pandas==1.2.2
Pillow==9.3.0
streamlit==1.17.0